Planets hips (#110)
* First draft of HiPS surveys support

This is based on the Toast survey support.  In fact I replaced the dss
button to use HiPS instead of toast, since I think eventually we should
only use HiPS instead of toast.

* Add a fading in effect to the hips tiles

Just like in toast.

* Hips: minor code factorisation

Preparing for the proper allsky support.

* Add function to create texture from a QImage

I need this for the hips manager code.

* Hips: use the allsky image if available

In order to avoid strange border effects, we split the allsky image into
small textures.  It's not very optimized, but should be good enough
since there are usually not too many tiles using allsky anyway.

* Hips: make survey url a Qt property

So that we can eventually change it from the ui.

* Hips: fix parsing format property

It can contain a list of values, in this case we iter them until we find
one format that we handle (only jpeg and png at the moment).

* Hips: render all parent tiles before children

That way we prevent losing the survey as we zoom in.

* Hips: use mipmaped textures

* Hips: supports galactic frame

* Hips: put properties into a QMap structure

That way we get them all directly, instead of having to add an argument
for each of them.

* HipsMgr get survey list from internet

Using the list at http://alaskybis.unistra.fr/hipslist.

* Add a combo to switch hips survey in view dialog

Not very useful as it is, but this will help to debug the code.

* StelHips: minor code cleanup

No need to use a connection for the network replies, since we can
directly use the 'isFinished' method.  Maked the code a bit easier to
follow.

* Remote texture supports https

* First test of planet hips surveys

This is just a test at the moment.  I try to use the same code than the
normal hips surveys, but we have to be careful about a few things:

- It's not trivial to compute the render order of the surveys, since the
  current formula assume that the survey cover the whole sky.  For the
  moment I fixed the value to 1.

- At low order, the sides of the planets is not smooth: I will need to
  make sure individual tiles get rendered using a grid.

- I might need to modify the planet shaders.

* Hips can split tile before render

That way we can render tiles of low levels with enough vertices so that
the planets borders look good.

Still experimental at the moment.

* Planet initShader can be called several times

Also makes it return the error status.

* StelHips: allow to pass a draw callback

So that we can use special shaders for the tiles.  I am not sure this is
the best approach though.  An other way would be to use an iterator.

* Planet hips survey uses planet shader

For the moment there is no support for rings.

* Planet HiPS survey better draw order computation

Taking into consideration the visible angle of the planet.

* Planet hips: support rings

The code is a bit tricky with the depth buffers.  I hope I don't break
other parts.

* Hips survey: render tiles on top of each other

We should try to minimize the number of tiles to render by skipping a
parent tile if all it's children are visible.

* Planet HiPS survey support moon shadows

The code is a bit ugly, because there are some matrix gymnastic in order
to get the planet rotation working.  I guess we could clean it up a bit.

* Planet HiPS fix oblatness

* Planet HiPS: support ring shadows

* Hips: fix non planet survey rendering

We need to make sure the orientation if correct for inside/outside
rendering.

Also the visibility test is different (though it would be better to use
a single generic test for all kind of surveys).

* Disable properties change debug outputs

Those are quite annoying, and they mess up the console output with large
properties (like json documents).

* Start to work on proper HiPS survey gui

I changed the Hips manager to load all the properties of the surveys
listed.  We need to do that in order to show the proper informations in
the ui, I guess we will need to use our own surveys server otherwise
alasky might not handle the high traffic.

* Hips survey ui show better informations

* Hips gui: allow to select which survey we render

* HipsMgr expose a list of HipsSurvey objects

Instead of exposing only the urls and properties.

It is simpler that way, and it will eventually allow to switch
individual surveys on and off.

* Hips: add a loading status to surveys

And show it in the UI.  For the moment I used text in the label.
It would probably be better to use some kind of icon for that.

* Fix compilation with Qt 5.4

* Hips: add some cache busting

Using an http argument: v=<hips_release_date>, we make sure that the
surveys can be cached as long as possible.

* Hips: fix int properties parsing

* Hips: show progress bars

I used one progress bar per survey, shown when we load the allsky or the
tiles.

* Hips: remove '[loading]' status in the ui

We don't need that now that we have the loading bars.

* Hips: allow to set the sources in the configs

I use an array like the satellites plugin does, so that we can have as
many sources as we want.

* StelTexture supports loading from file://

This is convenient sometime because Qt network requests only work with
proper urls.  The only thing is that I am not sure if the images get put
in the cache in that case (I hope not).

* Hips: support local surveys

* Auto assign planet HipsSurveys

Any loaded survey whose url filename matches a planet is automatically
assigned to the planet.

The way it works for the moment is that the HipsMgr emit a signal each
time a new survey is found.  The solasystem module listen to this signal
and assign the survey automatically.

Eventually I think we should add a property in the survey to specify if
we should use them on a given planet.  The HiPS spec doesn't cover this
use case.

It's a bit annoying that we have to include "StelHips.hpp" in the class
interface but I haven't found a way to pre-declare HipsSurveyP.

* Fix compilation with Qt 5.4

* Hips survey each have a visible property

So that we will eventually be able to render several at the same time.

I also added some logic for planet surveys, so that they are loaded from
the survey manager and visible in the survey list, but still rendered by
the planets.  Not sure if that is the best way to do it.

* Hips: group the surveys by types

For the moment we only consider deep sky and solar system surveys ('dss'
and 'sol').

I need to fix some of the data.stellarium.org surveys to specify the type
property.

* Fix compilation

* Hips: remove progressbar when a survey is disabled

* Don't show planet surveys by default

Since some of them are still not working properly.  For the moment we
have to enable them in the view dialog.

* Hips: add some doc comments

* DSS button is managed by the Hips survey manager

Instead of the toast manager.  The toast manager should probably
be removed at this point since we will probably only use HiPS from now
on.

I added a special attribute in the hips manager to toggle the DSS
survey, for the moment defined as the survey whose url ends with
DSSColor.

* Hips: fix bug with gui list update

For some reason if we try to update the list during the item changed
callback, it would select all the items in the list before the current
item.

I fix that by using a queued connection.