Change the meaning of the show DSS button
Previously I tried to keep the old behavior of actually showing the DSS
survey, but now that we have a surveys manager, it makes more sense to
use the button as a global toggle for it.

We probably need to add the toggle in the survey list window as well.