Fix for paths render with orthographic projection
This should fix bug #811.  The problem is that with the ortho projection
the points that are 'behind' the sphere should be hidden using culling,
but at the moment we don't really support that.  Here I add a clipping
test at the projection, even though this is not optimal (two clipped
points could still be linked by a partially visible line).

Hopefully this is enough to fix the bug.  I only activate this with the
ortho projection.