Updating Scripting Console (Fix #995)
Let's ask the file name in case, when user press button "Save script"
and file is new (it was not loaded by GUI or the text of script was
cleared) and overwrite him in other case. Now users can catch a "bugs"
like, "Aaaa! The file is overwtited and I lost my nice script!!!".