Added the Chang’e 5 lander location
A Chinese spacecraft successfully touched down on the moon Tuesday. The Chang’e 5 lander set down in the moon’s Oceanus Procellarum, or Ocean of Storms, region in the northern hemisphere of the near side of the moon, east of a volcanic plateau named Mons Rümker. The landing occurred at 10:11 a.m. EST (1511 GMT) Tuesday, according to the China National Space Administration. The space agency reported the lander touched down at 43.1 degrees north latitude and 51.8 degrees west longitude on the moon.