Added single asterism selection mode (fix #541, fix #921)
Added ability to limit the number of asterisms to draw based on selected asterisms. When the selected objects changed, check if some asterism are selected and display the matching asterisms if isolateAsterismSelected mode is activated. 

To avoid strange behavior of selection of asterisms the asterisms can be selected through Search Tool or scripts at the moment.