Fix bug in AstroCalc/Eclipses where type of local solar eclipses are incorrect (#2237)
* Fix bug in AstroCalc/Eclipses where type of local solar eclipses are incorrect

There is a bug in the AstroCalc's Local Solar Eclipses tool (in development) that must be fixed.

To reproduce
* open 'All Solar Eclipses' tab
* search for eclipses in the year 1986
* double click at 1986 Oct 3
* go to 'Local Solar Eclipses' tab
* search for eclipses in the same year

The result shows that 1986 Oct 3 was a partial eclipse as seen from the location of greatest eclipse along very narrow eclipse path (path width = 0.46 km) which is incorrect. For (unrealistically) perfect sphere Moon, this eclipse was a total eclipse with sub-second duration - 0.049 second - to be exact. In reality, it's beaded total eclipse. Both NASA web site and Stellarium tell us that it's hybrid eclipse - technically, it was total eclipse at mid point.

This PR will correctly classify this rare event as total eclipse as seen from the location and it will also fix similar situation on other dates when you are standing at the point of transition between total and annular eclipse.

Notes:
* This PR will fix the oblateness parameters of the Earth to be consistence across all related parts of Stellarium.
* As we reproduce the bug, we can see that it seems to be a slight difference of Sun & Moon positions at greatest eclipse point for some eclipses. This is still under investigations, would be addressed in a separate issue if I found a way to fix it.

Operating system: Windows 10
Graphics Card: Nvidia

* delete unused variable